move createStore to separate module
